Lint fix and reorganize files
