+++ src/interfaces/dropdownOptions.ts
@@ -1,0 +1,6 @@
+interface dropdownOptions {
+    value: string;
+    label: string;
+}
+
+export default dropdownOptions;
