+++ src/interfaces/navigationParameters.ts
@@ -1,0 +1,8 @@
+interface navigationParameters {
+    departureCRS: string;
+    arrivalCRS: string;
+    time: string;
+    timeWindow: string;
+}
+
+export default navigationParameters;
